+++ kitty/options/parse.py
@@ -1285,7 +1285,7 @@ class Parser:
         ans['text_composition_strategy'] = str(val)
 
     def text_fg_override_threshold(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
-        ans['text_fg_override_threshold'] = str(val)
+        ans['text_fg_override_threshold'] = float(val)
 
     def touch_scroll_multiplier(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['touch_scroll_multiplier'] = float(val)